requirements.txt
Flask==2.2.2
request==2.28.2
beautifulsoup4==4.12.0